app/src/requirements.txt
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0